requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.39.1
numpy==2.2.5
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==10.4.0
plotly==6.1.0
protobuf==4.25.7
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.0
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
streamlit-cropper==0.2.2
streamlit-plotly-events==0.0.6
tenacity==8.5.0
threadpoolctl==3.6.0
tifffile==2025.5.10
toml==0.10.2
tornado==6.5
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
validators==0.35.0
watchdog==5.0.3
zipp==3.21.0
